requirements.txt
streamlit==1.12.0
Pillow==8.0.1
opencv-python-headless==4.4.0.44
numpy==1.20.3
protobuf==3.20.3
